+++ packages/atomic/src/components/search/facets/atomic-segmented-facet/atomic-segmented-facet.tsx
@@ -17,7 +17,7 @@ import {
 } from '@coveo/headless';
 import {getFieldValueCaption} from '../../../../utils/field-utils';
 import {MapProp} from '../../../../utils/props-utils';
-import {FacetValuesGroup} from '../facet-values-group/facet-values-group';
+import {FacetValuesGroup} from '../../../common/facets/facet-values-group/facet-values-group';
 import {FacetSegmentedValue} from '../facet-segmented-value/facet-segmented-value';
 import {Hidden} from '../../../common/hidden';
 import {Bindings} from '../../atomic-search-interface/atomic-search-interface';